% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BayesRSamplerV2}
\alias{BayesRSamplerV2}
\title{BayesR sampler}
\usage{
BayesRSamplerV2(outputFile, seed, max_iterations, burn_in, thinning, X, Y,
  sigma0, v0E, s02E, v0G, s02G, cva)
}
\arguments{
\item{outputFile}{The file in which the samples aftare burnin will be stored}

\item{seed}{random seed}

\item{max_iterations}{total of number of samples taken}

\item{burn_in}{integer leq than max_iterations, number of samples used for burn in, after which, al samples will  be stored in the outputFile}

\item{thinning}{thinning regime, not implemented}

\item{X}{matrix of snp markers, or covariates of interest}

\item{Y}{vector of response variates, must have the same number of rows as X}

\item{sigma0}{variance of the zero-centered normal prior over the intercept}

\item{v0E}{degrees of  freedom of the prior inverse scaled chi-squared distribution over residues variance}

\item{s02E}{scale parameter of the prior inverse scaled chi-squared distribution over residues variance}

\item{v0G}{degrees of freedom of the prior inverse scaled chi-squared distribution over genetic effects variance}

\item{s02G}{scale parameter of the prior inverse scaled chi-squared distribution over residues variance}

\item{cva}{Vector of mixture components variances.}
}
\value{
(max_iteration - burn_in) modulo thinning samples saved in outputFile.
}
\description{
BayesR sampler
}
