% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BayesRSamplerV2Groups}
\alias{BayesRSamplerV2Groups}
\title{BayesRR sampler}
\usage{
BayesRSamplerV2Groups(
  outputFile,
  seed,
  max_iterations,
  burn_in,
  thinning,
  X,
  Y,
  sigma0,
  v0E,
  s02E,
  v0G,
  s02G,
  cva,
  groups,
  gAssign,
  fixed
)
}
\arguments{
\item{outputFile}{The file in which the samples aftare burnin will be stored}

\item{seed}{random seed}

\item{max_iterations}{total of number of samples taken.}

\item{burn_in}{integer leq than max_iterations, number of samples used for burn in, after which, al samples will  be stored in the outputFile}

\item{thinning}{thinning regime, not implemented}

\item{X}{matrix of snp markers, or covariates of interest}

\item{Y}{vector of response variates, must have the same number of rows as X}

\item{sigma0}{variance of the zero-centered normal prior over the intercept}

\item{v0E}{degrees of  freedom of the prior inverse scaled chi-squared distribution over residues variance}

\item{s02E}{scale parameter of the prior inverse scaled chi-squared distribution over residues variance}

\item{v0G}{degrees of freedom of the prior inverse scaled chi-squared distribution over genetic effects variance}

\item{s02G}{scale parameter of the prior inverse scaled chi-squared distribution over genetic effects variance}

\item{cva}{Matrix of mixture variances for groups, rows must be the same as number of groups, columns the same as number of mixtures.}

\item{groups}{number of groups}

\item{gAssign}{Vector of the same size as the number of columns of X, containing group assignments for each column(starting with group 0).}

\item{fixed}{Matrxi with same rows as X and Y , contains the fixed effects, those whose prior is always different than zero}
}
\description{
BayesRR sampler
}
