% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BRV2Grstart}
\alias{BRV2Grstart}
\title{BayesRR sampler function to re-start a preexisting BayesRR Markov chain.}
\usage{
BRV2Grstart(outputFile, seed, max_iterations, burn_in, thinning, mu, beta,
  sigmaE, sigmaGG, X, epsilon, components, sigma0, v0E, s02E, v0G, s02G,
  cva, groups, gAssign)
}
\arguments{
\item{outputFile}{The file in which the samples aftare burnin will be stored}

\item{seed}{random seed}

\item{max_iterations}{total of number of samples taken.}

\item{burn_in}{integer leq than max_iterations, number of samples used for burn in, after which, al samples will  be stored in the outputFile}

\item{thinning}{thinning regime, not implemented}

\item{mu}{last sample of mu from the previous chain.}

\item{beta}{vector of last sample of the effects (beta) from the last chain.}

\item{sigmaE}{value of residuals variance sigmaE in the last sample of the last chain.}

\item{sigmaGG}{vector of values for the group variances in the last sample in the last chain.}

\item{X}{matrix of snp markers, or covariates of interest}

\item{epsilon}{vecor of residuals from the last sample in the last chain.}

\item{components}{vector of mixture assignments for each effect in the last sample for the last chain.}

\item{sigma0}{variance of the zero-centered normal prior over the intercept}

\item{v0E}{degrees of  freedom of the prior inverse scaled chi-squared distribution over residues variance}

\item{s02E}{scale parameter of the prior inverse scaled chi-squared distribution over residues variance}

\item{v0G}{degrees of freedom of the prior inverse scaled chi-squared distribution over genetic effects variance}

\item{s02G}{scale parameter of the prior inverse scaled chi-squared distribution over genetic effects variance}

\item{cva}{Matrix of mixture variances for groups, rows must be the same as number of groups, columns the same as number of mixtures.}

\item{groups}{number of groups}

\item{gAssign}{Vector of the same size as the number of columns of X, containing group assignments for each column(starting with group 0).}
}
\description{
BayesRR sampler function to re-start a preexisting BayesRR Markov chain.
}
